Revert channel to unlisted when signing xpi
The listed channel does not return until the extension is
approved by Firefox Add-ons review, which can take days.

Instead, change the version number to an unlisted-specific
one by appending -git to the version number.
Unlisted version cannot be reused when uploading to Firefox Add-ons,
which is why it needs a separate version number.